requirements.txt
# LLM Provider - LiteLLM with Groq-only support
litellm==1.57.5
tenacity>=8.2.0  # Retry logic for LLM API calls
mcp
python-dotenv
PyPDF2
pycryptodome  # For encrypted PDFs
langchain
chromadb
pydantic-settings
fastapi
uvicorn[standard]
tiktoken  # For token-based text chunking
sentence-transformers  # Local embeddings for Groq-only workflows

# Web scraping and automation
beautifulsoup4>=4.12.0
requests>=2.31.0
schedule>=1.2.0
aiohttp>=3.9.0  # Async HTTP for PubMed API
nest-asyncio>=1.6.0  # Allow nested asyncio loops (for PMC fetcher in FastAPI)

# PubMed/Bioinformatics
biopython>=1.83  # PubMed article parsing utilities

# Data analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Future LLM provider support (optional)
# openai  # Uncomment to enable OpenAI provider
# anthropic  # Uncomment to enable Anthropic/Claude provider
# ollama  # Uncomment to enable local Ollama models
